+++ src/Fantomas.Core.Tests/InterpolatedStringTests.fs
@@ -327,8 +327,8 @@ $"
                          let b = y
                          let c = 9
                          foo ()
-                     with
-                     | ex -> interpolationFailed ()}
+                     with ex ->
+                         interpolationFailed ()}
 "
 """